+++ src/main/java/com0/coastalcasa/Mapper/ListingMapper.java
@@ -27,8 +27,8 @@ public interface ListingMapper {
     @Select("SELECT * FROM listing WHERE landlord_email = #{landlord_email}")
     List<Listing> findByLandlordEmail(String landlordEmail);
     
-    @Insert("INSERT INTO listing(landlord_email, description, location, price, num_bathrooms, num_bedrooms, amenities) " +
-        "VALUES (#{landlord_email}, #{description}, #{location}, #{price}, #{num_bathrooms}, #{num_bedrooms}, #{amenities})")
+    @Insert("INSERT INTO listing(landlord_email, name, description, location, price, num_bathrooms, num_bedrooms, amenities) " +
+        "VALUES (#{landlord_email},#{name}, #{description}, #{location}, #{price}, #{num_bathrooms}, #{num_bedrooms}, #{amenities})")
     @Options(useGeneratedKeys = true, keyProperty = "id", keyColumn = "id")
     int insert(Listing listing);
 
@@ -42,6 +42,7 @@ public interface ListingMapper {
             "SELECT * FROM listing ",
             "WHERE 1 = 1 " ,
             "<if test='landlordEmail != null'> AND landlord_email = #{landlordEmail} </if>",
+            "<if test='name != null'> AND name = #{name} </if>",
             "<if test='description != null'> AND description = #{description} </if>" ,
             "<if test='location != null'> AND location LIKE #{location} </if>" ,
             "<if test='price != -1.0'> AND price = #{price} </if>" ,
@@ -49,7 +50,7 @@ public interface ListingMapper {
             "<if test='num_bedrooms != -1'> AND num_bedrooms = #{num_bedrooms} </if>" ,
             "<if test='amenities != null'> AND amenities LIKE #{amenities} </if>" ,
             "</script>"})
-    List<Listing> searchListings(@Param("landlordEmail") String landlordEmail, @Param("description") String description,@Param("location") String location, @Param("price") double price, @Param("num_bathrooms") int num_bathrooms, @Param("num_bedrooms") int num_bedrooms, @Param("amenities") String amenities);
+    List<Listing> searchListings(@Param("name") String name,@Param("landlordEmail") String landlordEmail, @Param("description") String description,@Param("location") String location, @Param("price") double price, @Param("num_bathrooms") int num_bathrooms, @Param("num_bedrooms") int num_bedrooms, @Param("amenities") String amenities);
     
     @Update("UPDATE listing SET landlord_email = #{landlord_email}, description = #{description}, location = #{location}, " +
             "price = #{price}, num_bedrooms = #{num_bedrooms}, num_bathrooms = #{num_bathrooms}, amenities = #{amenities} " +
