+++ src/main/java/com0/coastalcasa/Security/WebMvcConfig.java
@@ -19,6 +19,7 @@ public class WebMvcConfig implements WebMvcConfigurer {
                 .addPathPatterns("/listings/test")
                 .addPathPatterns("/listings/update/**")
                 .addPathPatterns("/listings/delete/**")
+                .addPathPatterns("/applicants/delete/**")
                 .excludePathPatterns("/landlord/**")
                 .excludePathPatterns("/landlord/**");
     }
