+++ src/main/resources/application.properties
@@ -1,3 +1,3 @@
 # change this to dev, but do not commit this file
 # also change the password and username of your DB defined in application-dev.properties file
-spring.profiles.active=prod
+spring.profiles.active=dev
