+++ src/main/resources/scripts/createDB.sql
@@ -26,6 +26,7 @@ CREATE TABLE IF NOT EXISTS landlord (
 CREATE TABLE IF NOT EXISTS listing (
   id INT PRIMARY KEY AUTO_INCREMENT,
   landlord_email VARCHAR(255) NOT NULL,
+  name TEXT NOT NULL,
   description TEXT NOT NULL,
   location VARCHAR(255) NOT NULL,
   price DECIMAL(10, 2) NOT NULL,
